Add initial implementation of search app bar and product screen
This commit adds the initial implementation of the search app bar and product screen. It includes the addition of new font and image assets, as well as the creation of the `HomeAppBar` widget and the `ProductsScreen` widget. The `MyApp` widget has also been updated to include the `ScreenUtilInit` widget for screen size adaptation. The changes made are necessary for the development of the product feature.